[coursework-ris] ??????? ??????? ???????. ????? ????: 1. ? AbstractFilter ? ?????? ????? ??? Long, ? ? ????? ??????? ???? ???????? ??????????????? ???? ?? Integer 2. ? checkAgreement ???? ???????? ???????? ????????, ?? ???????? ?? ???????? ?? item.getText(column), ??????? ?????? ?????????? String

+++ courseworkRIS/src/main/java/org/courseworks/ris/mappings/hprepair/Cars.java
@@ -14,7 +14,7 @@ public class Cars extends AbstractEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    public long id;
+    public Integer id;
 
     public String Name;
